Stop paying for the AI stack on installs that opted out of it
smartgallery_ai.faces imported torch at module scope for one CUDA
function, so every process that touched the gallery paid 1.7s and
several hundred megabytes -- including installs running with
ENABLE_AI_DAM=false, which never reach that code. Importing it where it
is used takes the module import from 2.67s to 0.94s and makes the
ImportError the function's own docstring promises reachable, so a
missing torch falls back to the faiss and numpy backends instead of
failing the package.

import_faiss() had the same shape as a bug: `import faiss` inside the
vendored-GPU branch makes the name local to the whole function, so the
`return faiss` at the end raised UnboundLocalError on every path that
skipped that branch -- all of Linux and macOS, any Windows box without
nvidia-smi, and anyone setting AI_DAM_FAISS_GPU=0. The documented
faiss-cpu fallback had never once run.

Scans below PARALLEL_SCAN_MIN_FILES now stay in this process. A pool
worker has to import this module before it can look at a picture, which
costs about a second whether the batch is three files or three thousand,
and dropping a handful of new files in is the common case.

+++ benchmarks/face_detection_recall.py
@@ -23,6 +23,10 @@ import os
 import sys
 import time
 
+from PIL import Image
+
+from smartgallery_ai import AIConfig
+from smartgallery_ai.faces import get_face_backend
 
 
 def load_repo():
@@ -81,7 +85,6 @@ def collect_ground_truth(backend, root, limit):
     native-res policy already handles — exactly the bias the >=300 band
     exists to expose.
     """
-    from PIL import Image
 
     sources = []
     paths = []
@@ -112,10 +115,7 @@ def collect_ground_truth(backend, root, limit):
 
 def main() -> None:
     load_repo()
-    from PIL import Image
 
-    from smartgallery_ai import AIConfig
-    from smartgallery_ai.faces import get_face_backend
 
     ap = argparse.ArgumentParser(description=__doc__)
     ap.add_argument("--dir", required=True, help="image corpus root")
@@ -167,8 +167,8 @@ def main() -> None:
                 target = int(tag[len("scale->"):-2])
                 scale = target / min(gt_boxes[0][2], gt_boxes[0][3])
             if scale != 1.0:
-                nw = max(1, int(round(img.size[0] * scale)))
-                nh = max(1, int(round(img.size[1] * scale)))
+                nw = max(1, round(img.size[0] * scale))
+                nh = max(1, round(img.size[1] * scale))
                 img = img.resize((nw, nh), Image.LANCZOS)
             gts = [tuple(v * scale for v in b) for b in gt_boxes]
             t0 = time.perf_counter()
